Added: A bit of game.js. Most of role integration. Currently need to pass thisRoom every time we use a role's function.
